# Added sessionTracer to log graph operations.