Still missing a test and a benchmark.

@silky, this should be better than the patch I previously suggested.