@blackgnezdo 2016-11-14
I tried this change with #27 benchmark and it detected no differences with logging being set to the default no-op.

BEFORE
% git status
# On branch feed_fetch_bench
% stack build --bench :FeedFetchBench
...
benchmarking feedFetch/4 byte
time                 54.85 μs   (53.90 μs .. 55.71 μs)
                     0.998 R²   (0.996 R² .. 0.999 R²)
mean                 55.44 μs   (54.87 μs .. 56.23 μs)
std dev              2.375 μs   (1.823 μs .. 3.470 μs)
variance introduced by outliers: 47% (moderately inflated)

benchmarking feedFetch/4 KiB
time                 90.85 μs   (88.84 μs .. 92.30 μs)
                     0.997 R²   (0.994 R² .. 0.999 R²)
mean                 91.43 μs   (89.95 μs .. 92.82 μs)
std dev              4.722 μs   (3.419 μs .. 6.387 μs)
variance introduced by outliers: 54% (severely inflated)

benchmarking feedFetch/4 MiB
time                 129.9 ms   (127.4 ms .. 133.4 ms)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 130.1 ms   (129.5 ms .. 131.4 ms)
std dev              1.152 ms   (396.4 μs .. 1.674 ms)
variance introduced by outliers: 11% (moderately inflated)

Benchmark FeedFetchBench: FINISH

AFTER
% git status
# On branch tracing-bench
% stack build --bench :FeedFetchBench
...
benchmarking feedFetch/4 byte
time                 54.35 μs   (53.44 μs .. 55.10 μs)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 53.24 μs   (52.45 μs .. 53.87 μs)
std dev              2.544 μs   (2.095 μs .. 3.125 μs)
variance introduced by outliers: 52% (severely inflated)

benchmarking feedFetch/4 KiB
time                 91.40 μs   (90.46 μs .. 92.33 μs)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 92.37 μs   (91.72 μs .. 93.09 μs)
std dev              2.295 μs   (1.819 μs .. 3.426 μs)
variance introduced by outliers: 22% (moderately inflated)

benchmarking feedFetch/4 MiB
time                 131.5 ms   (130.2 ms .. 133.0 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 131.9 ms   (131.6 ms .. 132.2 ms)
std dev              464.8 μs   (335.3 μs .. 640.1 μs)
variance introduced by outliers: 11% (moderately inflated)