@blackgnezdo 2016-11-14
I also tried with a trace statement inserted before FFI.run. It was similarly unaffected.